+++ fast_rx/lib/src/rx/composite/rx_composite.dart
@@ -11,6 +11,9 @@ mixin RxCompositeMixin<T> on Rx<T> {
   @protected
   void addRx(Rx rx) {
     if (_subs[identityHashCode(rx)] != null) return;
+
+    // Pana was complaining about this for no reason
+    // ignore: cancel_subscriptions
     final sub = rx.stream.listen((_) => notify());
     _subs[identityHashCode(rx)] = sub;
   }